🚧 VN1FN4 task: keep recovery tests within file-size limits
